add better deletion handling, operations empty state and remove unwanted tabIndex

+++ src/app/(protected)/operations/components/Operation.tsx
@@ -1,4 +1,16 @@
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from '@/components/ui/alert-dialog';
 import { Button } from '@/components/ui/button';
+import { handleShiftClick } from '@/lib/handleShiftClick';
 import { cn } from '@/lib/utils';
 import { stockProfileQuery } from '@/queries/stockProfileQuery';
 import { useConvexMutation } from '@convex-dev/react-query';
@@ -74,6 +86,8 @@ export function OperationActions({
 }: {
   operation: {
     _id: Id<'operations'>;
+    symbol: string;
+    quantity: number;
   };
 }) {
   const { mutate: deleteOperation } = useMutation({
@@ -82,14 +96,38 @@ export function OperationActions({
 
   return (
     <div className="flex flex-row gap-2 items-center">
-      <Button
-        variant="ghost"
-        size="icon"
-        className="cursor-pointer"
-        onClick={() => deleteOperation({ id: operation._id })}
-      >
-        <Trash2Icon />
-      </Button>
+      <AlertDialog>
+        <AlertDialogTrigger asChild>
+          <Button
+            variant="ghost"
+            size="icon"
+            className="cursor-pointer"
+            onClick={handleShiftClick({ onShiftClick: () => deleteOperation({ id: operation._id }) })}
+          >
+            <Trash2Icon />
+          </Button>
+        </AlertDialogTrigger>
+        <AlertDialogContent>
+          <AlertDialogHeader>
+            <AlertDialogTitle>
+              Eliminar {operation.quantity} acciones de {operation.symbol}
+            </AlertDialogTitle>
+            <AlertDialogDescription>¿Estás seguro de querer eliminar esta operación?</AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel className="cursor-pointer" type="button">
+              Cancelar
+            </AlertDialogCancel>
+            <AlertDialogAction
+              className="cursor-pointer bg-red-500 text-white hover:bg-red-600"
+              type="submit"
+              onClick={() => deleteOperation({ id: operation._id })}
+            >
+              Eliminar
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
     </div>
   );
 }
